// src/types/patient.ts
export interface DiagnosisHistory {
  month: string
  year: number
  blood_pressure: {
    systolic: {
      value: number
      levels: string
    }
    diastolic: {
      value: number
      levels: string
    }
  }
  heart_rate: {
    value: number
    levels: string
  }
  respiratory_rate: {
    value: number
    levels: string
  }
  temperature: {
    value: number
    levels: string
  }
}

export interface DiagnosticListItem {
  name: string
  description: string
  status: string
}

export interface LabResult {
  name: string
}

export interface Patient {
  name: string
  gender: string
  age: number
  profile_picture: string
  date_of_birth: string
  phone_number: string
  emergency_contact: string
  insurance_type: string
  diagnosis_history: DiagnosisHistory[]
  diagnostic_list: DiagnosticListItem[]
  lab_results: LabResult[]
}